Remove git from the bundler process
