+++ index.js
@@ -3,9 +3,7 @@
 const fs = require('fs');
 const path = require('path');
 const Listr = require('listr');
-const execa = require('execa');
 const chalk = require('chalk');
-const del = require('del');
 const UpdaterRenderer = require('listr-update-renderer');
 const moment = require('moment');
 const argv = require('minimist')(process.argv.slice(2), {
@@ -29,213 +27,15 @@ const argv = require('minimist')(process.argv.slice(2), {
 const ENV_FILE = fs.existsSync('.env') ? '.env' : 'env/.env';
 require('dotenv').config({ path: ENV_FILE });
 
-const { debug, success, error, about, warn } = require('./lib/helpers/log')('proton-bundler');
-const extractArgument = require('./lib/helpers/arguments');
-const coucou = require('./lib/helpers/coucou');
-const { bash, script } = require('./lib/helpers/cli');
-const {
-    customBundler: { tasks: customTasks, config: customConfig },
-    getCustomHooks
-} = require('./lib/custom');
-const { pull, push, getConfig, generateChangelog } = require('./lib/git');
+const { debug, success, error } = require('./lib/helpers/log')('proton-bundler');
 const askDeploy = require('./lib/ask');
-const buildRemote = require('./lib/buildRemote');
+const { get: getConfig } = require('./lib/config');
+const getTasks = require('./lib/tasks/index');
+const remoteBuildProcesss = require('./lib/tasks/remote');
+const changelogProcess = require('./lib/tasks/changelog');
+const { script } = require('./lib/helpers/cli');
 
 const PKG = require(path.join(process.cwd(), 'package.json'));
-
-const getTasks = (branch, { isCI, flowType = 'single', forceI18n, appMode, runI18n, isRemoteBuild, featureFlags }) => {
-    const { EXTERNAL_FILES = ['.htaccess'] } = customConfig;
-    const { hookPreTasks, hookPostTasks, hookPostTaskClone, hookPostTaskBuild, customConfigSetup } = getCustomHooks(
-        customTasks({
-            branch,
-            isCI,
-            flowType,
-            appMode,
-            forceI18n,
-            runI18n,
-            isRemoteBuild,
-            featureFlags
-        })
-    );
-
-    const configTasks = customConfigSetup.length
-        ? customConfigSetup
-        : [
-              {
-                  title: 'Setup app config',
-                  task() {
-                      return bash('npx proton-pack', process.argv.slice(2));
-                  }
-              }
-          ];
-
-    const list = [
-        {
-            title: 'Check if the deploy branch exists',
-            skip() {
-                if (!(process.env.QA_BRANCHES || '').length) {
-                    return 'If you do not have QA_BRANCHES inside your .env you cannot deploy to something new';
-                }
-
-                if (`deploy-${process.env.QA_BRANCH}` === branch) {
-                    return '🤖 No need to check for this branch, it exists';
-                }
-
-                const branches = process.env.QA_BRANCHES.split(',').join('|');
-                // Do not try to deploy on QA or cobalt
-                if (new RegExp(`^deploy-(cobalt|${branches})$`).test(branch)) {
-                    return '✋ You shall not deploy to QA';
-                }
-            },
-            enabled: () => !isCI && !/dev|beta|prod|tor|old/.test(branch),
-            async task() {
-                // For the CI to force SSH
-                if (process.env.GIT_REMOTE_URL_CI && argv.fromCi) {
-                    await bash(`git remote set-url origin ${process.env.GIT_REMOTE_URL_CI}`);
-                }
-                return script('createNewDeployBranch.sh', ['--check', branch.replace('deploy-', '')], 'inherit');
-            }
-        },
-        ...hookPreTasks,
-        {
-            title: 'Save dependencies if we need',
-            enabled: () => !isRemoteBuild && !isCI && /dev|beta|alpha/.test(branch),
-            task() {
-                return script('updatePackageLock.sh', [argv['default-branch']]);
-            }
-        },
-        {
-            title: 'Clear previous dist',
-            async task() {
-                await del(['dist', 'distCurrent', 'distback'], { dryRun: false });
-                isCI && (await bash('mkdir dist'));
-            }
-        },
-        {
-            title: 'Prepare build missing dir',
-            async task() {
-                await bash("[ ! -d 'src/i18n' ] && mkdir src/i18n || echo 'Already there'");
-            }
-        },
-        {
-            title: 'Lint sources',
-            enabled: () => argv.lint !== false && !isRemoteBuild,
-            task: () => execa('npm', ['run', 'lint'])
-        },
-        ...configTasks,
-        {
-            title: 'Extract git env for the bundle',
-            async task(ctx) {
-                const { commit, branch, tag } = await getConfig();
-                ctx.originCommit = commit;
-                ctx.originBranch = branch;
-                ctx.tag = tag;
-                debug(ctx, 'git env bundle');
-            }
-        },
-        {
-            title: `Pull dist branch ${branch}`,
-            enabled: () => !isCI,
-            task: () => pull(branch, argv.forceFetch, argv.fromCi)
-        },
-        ...hookPostTaskClone,
-        {
-            title: 'Copy some files',
-            task() {
-                const rule = EXTERNAL_FILES.length > 1 ? `{${EXTERNAL_FILES.join(',')}}` : EXTERNAL_FILES.join(',');
-                return bash(`cp src/${rule} dist/`);
-            }
-        },
-        {
-            title: 'Upgrade translations inside the app',
-            enabled: () => forceI18n || (runI18n && !isCI && /prod|beta/.test(branch)),
-            task() {
-                return execa('npm', ['run', 'i18n:getlatest']);
-            }
-        },
-        {
-            title: 'Build the application',
-            async task(ctx = {}) {
-                const args = process.argv.slice(2);
-                if (appMode === 'standalone') {
-                    const output = await bash('npm', ['run', 'build:standalone', '--', ...args]);
-                    ctx.outputBuild = output;
-                    return true;
-                }
-
-                const output = await bash('npm', ['run', 'build', '--', ...args]);
-                ctx.outputBuild = output;
-                return true;
-            }
-        },
-        {
-            title: 'Check the build output content',
-            // Extract stdout from the output as webpack can throw error and still use stdout + exit code 0
-            async task(ctx = {}) {
-                await script('validateBuild.sh');
-                delete ctx.outputBuild; // clean as we won't need it anymore
-            }
-        },
-        ...hookPostTaskBuild,
-        {
-            title: 'Generate the version info',
-            task(ctx) {
-                const { tag = `v${PKG.version}`, originCommit, originBranch } = ctx || {};
-                const fileName = path.join('dist', 'assets/version.json');
-                const version = PKG['version-beta'] || tag; // custom version for v4
-
-                const args = [
-                    `--tag ${version}`,
-                    `--commit ${originCommit}`,
-                    `--branch ${originBranch}`,
-                    `--output ${fileName}`,
-                    '--debug'
-                ];
-
-                return script('createVersionJSON.sh', args);
-            }
-        },
-        {
-            title: `Push dist to ${branch}`,
-            enabled: () => !isCI,
-            task: (ctx) => push(branch, ctx)
-        },
-        {
-            title: 'Update crowdin with latest translations',
-            enabled: () => runI18n && !isCI && /prod|beta/.test(branch),
-            task() {
-                return execa('npm', ['run', 'i18n:upgrade']);
-            }
-        },
-        ...hookPostTasks,
-        {
-            title: 'Inform us about new translations available',
-            enabled: () => runI18n && !isCI && /prod|beta/.test(branch),
-            task() {
-                const [, type] = branch.match(/deploy-(\w+)/) || [];
-                return execa('npx', ['proton-i18n', 'coucou', type]);
-            }
-        }
-    ];
-    return list;
-};
-
-/**
- * Get the API dest when we deploy.
- * You can use the custom config from proton.bundler.js to get it
- * @return {Promise<String>}
- */
-async function getAPIUrl() {
-    if (customConfig.apiUrl) {
-        return customConfig.apiUrl;
-    }
-
-    const args = process.argv.slice(2);
-    const { stdout } = await bash('npx proton-pack print-config', args);
-    debug(stdout, 'print-config output');
-    const [, url] = stdout.match(/apiUrl": "(.+)"(,*?)/);
-    return url;
-}
 
 async function main() {
     /*
@@ -246,49 +46,18 @@ async function main() {
         So let's put an end to the current deploy.
      */
     if (argv.remote) {
-        await buildRemote(PKG);
-        const args = process.argv.slice(2).filter((key) => !/--remote/.test(key));
-        return script('builder.sh', [PKG.name, ...args], 'inherit'); // inherit for the colors ;)
+        return remoteBuildProcesss(PKG);
     }
 
-    // Custom local deploy for the CI
-    const isCI = process.env.NODE_ENV_DIST === 'bundle-only';
-    const branch = extractArgument(argv.branch) || '';
-    const featureFlags = extractArgument(argv.featureFlags) || '';
-    const flowType = argv.flow;
-    const runI18n = argv.i18n;
-    const forceI18n = argv.localize;
-    const appMode = argv.appMode;
-    const isRemoteBuild = argv.source === 'remote';
-
-    debug({ customConfig, argv, branch }, 'configuration deploy');
-
-    if (!branch && !isCI) {
-        throw new Error('You must define a branch name. --branch=XXX');
-    }
-
-    const apiUrl = await getAPIUrl();
-
-    process.env.NODE_ENV_BRANCH = branch;
-    process.env.NODE_ENV_API = apiUrl;
-
-    about({
-        ...(!isCI && { branch }),
-        apiUrl,
-        appMode,
-        isRemoteBuild,
-        featureFlags,
-        isCI,
-        SENTRY: process.env.NODE_ENV_SENTRY
-    });
-
+    const config = await getConfig(argv);
     const start = moment(Date.now());
-    const listTasks = getTasks(branch, { isCI, flowType, forceI18n, appMode, runI18n, isRemoteBuild, featureFlags });
+    const listTasks = getTasks({ config, PKG, argv });
     const tasks = new Listr(listTasks, {
         renderer: UpdaterRenderer,
         collapse: false
     });
 
+    const { isCI, branch } = config;
     await tasks.run();
 
     const now = moment(Date.now());
@@ -321,25 +90,7 @@ if (argv._.includes('log-commits')) {
 }
 
 if (argv._.includes('changelog')) {
-    const { branch, mode, api } = argv;
-    const url = argv.url || (PKG.bugs || {}).url;
-    const isV4 = mode === 'v4';
-    debug({ argv }, 'arguments');
-
-    if (!['dev', 'v4', process.env.QA_BRANCH].includes(branch) && !isV4) {
-        return warn(`No changelog available for the branch ${branch}`); // not available
-    }
-
-    if (!url && !isV4) {
-        return warn('No URL found for the issues');
-    }
-
-    return generateChangelog(branch, url, isV4).then((data) => {
-        const env = isV4 ? 'https://v4.protonmail.blue' : branch;
-        if (data) {
-            coucou.send(data, { env, mode: 'changelog', api }, PKG);
-        }
-    });
+    return changelogProcess(argv, PKG);
 }
 
 main().catch(error);
